Update v0.9 - battleship is now multiplayer ready!*
updated transaction content data recognition in order to exclude all
transactions associated with a given targetAddress or fleetAddress that
are sent before the current instance of the game has started

finalized transaction content data recognition to only recognize the
latest incoming transaction and update your fleet table accordingly

extended displayIncoming function with hit scan detection based on your
own fleet position and incoming projectile coordinates

updated the saveConnectionSettings function to allow for multiplayer

*: This instance of the program cannot yet update your enemy fleet table
to reflect whether or not one of your projectiles has either struck an
enemy ship or dropped into the ocean. An active communication layer
with; as well as trust in your opponent is highly recommended until the
feature is implemented!